pyproject.toml
[tool.poetry]
name = "dillagent"
version = "1.0.7"
description = "Agentic LLM library"
authors = ["Dillan Khurana <buzzoo123@users.noreply.github.com>"]
license = "LICENSE"
readme = "README.md"
homepage = "https://github.com/buzzoo123/dillagent"
repository = "https://github.com/buzzoo123/dillagent"

[tool.poetry.dependencies]
python = "^3.11"
annotated-types = "0.6.0"
anthropic = "0.19.2"
anyio = "4.3.0"
beautifulsoup4 = "4.12.3"
certifi = "2024.2.2"
charset-normalizer = "3.3.2"
colorama = "0.4.6"
distro = "1.9.0"
docutils = "0.20.1"
exceptiongroup = "1.2.0"
filelock = "3.13.1"
fsspec = "2024.2.0"
googlesearch-python = "1.2.3"
h11 = "0.14.0"
httpcore = "1.0.4"
httpx = "0.27.0"
huggingface-hub = "0.21.4"
idna = "3.6"
importlib-metadata = "7.0.2"
jaraco-classes = "3.3.1"
keyring = "24.3.1"
markdown-it-py = "3.0.0"
mdurl = "0.1.2"
more-itertools = "10.2.0"
nh3 = "0.2.15"
openai = "1.13.3"
packaging = "24.0"
pkginfo = "1.10.0"
pydantic = "2.6.3"
pydantic-core = "2.16.3"
pygments = "2.17.2"
python-dotenv = "1.0.1"
pywin32-ctypes = "0.2.2"
pyyaml = "6.0.1"
readme-renderer = "43.0"
requests = "2.31.0"
requests-toolbelt = "1.0.0"
rfc3986 = "2.0.0"
rich = "13.7.1"
sniffio = "1.3.1"
soupsieve = "2.5"
termcolor = "2.4.0"
tokenizers = "0.15.2"
tqdm = "4.66.2"
twine = "5.0.0"
typing-extensions = "4.10.0"
urllib3 = "2.2.1"
zipp = "3.18.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
